fixed prompting to include asserts

+++ automodel.py
@@ -7,7 +7,7 @@ from transformers import AutoModelForCausalLM, AutoTokenizer
 
 class Model:
     def __init__(self, name, revision, tokenizer_name=None, tokenizer_revision=None):
-        self.model = AutoModelForCausalLM.from_pretrained(name, revision=revision, torch_dtype=torch.float16, trust_remote_code=True).cuda()
+        self.model = AutoModelForCausalLM.from_pretrained(name, revision=revision, torch_dtype=torch.bfloat16, trust_remote_code=True).cuda()
         self.tokenizer = AutoTokenizer.from_pretrained(tokenizer_name or name, revision=tokenizer_revision or revision, padding_side="left", trust_remote_code=True)
         self.tokenizer.pad_token = "<|endoftext|>"
